requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
openpyxl==3.1.5
pydantic==2.9.2
python-multipart==0.0.12
python-docx==1.1.2
httpx==0.27.0
reportlab==4.2.5
pyyaml==6.0.2
weasyprint>=68.0
Jinja2>=3.1
pytest>=8.0
